refactor(tokenizer): widen instruction length
Some gameboy instructions are sadly 3 bytes long rather than the max 2 I assumed.

+++ include/libbeemu/device/processor/tokenizer.h
@@ -15,10 +15,12 @@ extern "C"
 	/**
 	 * @brief Tokenize an instruction.
 	 *
-	 * @param instruction Instruction to tokenize
+	 * @param instruction Instruction to tokenize, instructions are considered
+	 * right padded, with the leftmost byte always being 0x00 and the latter
+	 * three bytes possibly being important.
 	 * @return BeemuInstruction Instrcution, tokenized.
 	 */
-	BeemuInstruction *beemu_tokenizer_tokenize(uint16_t instruction);
+	BeemuInstruction *beemu_tokenizer_tokenize(uint32_t instruction);
 
 	/**
 	 * @brief Free a instruction token.
